+++ lsp/src/main/kotlin/com/falsepattern/zigbrains/lsp/settings/ZLSSettingsPanel.kt
@@ -279,7 +279,7 @@ class ZLSSettingsPanel(private val project: Project) : ZigProjectConfigurationPr
         }
 
     private fun dispatchAutodetect(force: Boolean) {
-        project.zigCoroutineScope.launchWithEDT {
+        project.zigCoroutineScope.launchWithEDT(ModalityState.any()) {
             withModalProgress(ModalTaskOwner.component(zlsPath), "Detecting ZLS...", TaskCancellation.cancellable()) {
                 autodetect(force)
             }